adk_cycling/requirements.txt
google-adk
google-cloud-bigquery
fastapi
uvicorn[standard]
python-multipart
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
google-api-python-client
itsdangerous
httpx
jinja2
pywebpush
